backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pillow==11.0.0
numpy==2.2.1
httpx==0.28.1
python-whois==0.9.27
imagehash==4.3.1
playwright==1.50.0
aiosqlite==0.20.0
onnxruntime==1.20.1